requirements.txt
googlesearch-python
requests
beautifulsoup4
tldextract
